Problem: incorrect geometry is being cropped
The user draws regions on a scaled-down
image. The geometry was not being scaled
up in pipes and xml, leading to incorrect
crops.

Solution: correct geometry for pipes and in
sequence xml. Region geometry is still the
native ui/scaled geometry, this makes reloads
render correctly, but means that using region
geometry as-is will be problematic. This should
be changed.